chore: add middleware
