+++ backend/src/index.js
@@ -3,10 +3,12 @@ const { ApolloServerPluginDrainHttpServer } = require("apollo-server-core");
 const express = require("express");
 const http = require("http");
 const Mongoose = require("mongoose");
-const jwt = require("jsonwebtoken");
 
 require("dotenv").config();
 const { typeDefs, resolvers, datasources } = require("./graphql");
+
+// middleware
+const middleware = require("./middleware/index");
 
 // set up data sources
 const dataSources = () => ({
@@ -38,32 +40,7 @@ async function startApolloSever() {
     resolvers,
     dataSources,
     context: ({ req }) => {
-      const header = req.headers.authorization;
-      let auth = {
-        isAuth: false,
-      };
-      if (!header) {
-        return { auth };
-      }
-
-      const token = header.split(" ");
-      if (!token) {
-        return { auth };
-      }
-
-      let decodeToken;
-
-      try {
-        decodeToken = jwt.verify(token[1], process.env.JWT_SECRET);
-      } catch (err) {
-        return { auth };
-      }
-
-      auth = {
-        isAuth: true,
-        ...decodeToken,
-      };
-      return { auth };
+      return middleware(req);
     },
     plugins: [ApolloServerPluginDrainHttpServer({ httpServer })],
   });
