+++ backend/src/middleware/index.js
@@ -1,0 +1,10 @@
+const { auth, isAuthenticated } = require("./auth");
+
+module.exports = (req) => {
+  req.user = auth(req);
+
+  return {
+    req,
+    isAuthenticated,
+  };
+};
